refactor(templates): extract sandbox-aware reader for provider rule reuse
Closes #727
