+++ directory-tree.md
@@ -196,10 +196,10 @@ maestro/
 │   │   ├── error.rs                       # `TemplateError` enum (thiserror); variants: UnknownPlaceholder, InvalidPlaceholder, SandboxEscape, IncludeCycle, ManifestMissing, ManifestParse, FileMissing, Io, UnterminatedPlaceholder, UnsupportedByProvider, UnsupportedManifestVersion  [Issue #701]
 │   │   ├── manifest.rs                    # `Manifest::load(path)` parses `.maestro/templates/manifest.toml` into typed structs; `#[serde(deny_unknown_fields)]` on all TOML tables  [Issue #701]
 │   │   ├── provider_rules/
-│   │   │   ├── mod.rs                     # `TemplateProviderRules` trait (per-provider placeholder validation + unsupported-feature guard); `NullRules` fail-closed stub; `null_rules()` free fn returning `&'static NullRules`; re-exports `ClaudeRules`, `codex_rules()`, and `http_generic_rules()`  [Issue #701, #703, #704, #705]
-│   │   │   ├── claude.rs                  # `ClaudeRules` — concrete `TemplateProviderRules` impl for Claude Code; sandbox-validated `include()` guard; 11 unit tests  [Issue #703]
-│   │   │   ├── codex.rs                   # `CodexRules` — concrete `TemplateProviderRules` impl for Codex; `target_dir()=None`; `invoke_subagent` inlines as `## Sub-task`; `hook_gate` emits provider-neutral bash path; `skill_link` inlines full body; `subagent_list` returns link-free Markdown table  [Issue #704]
-│   │   │   └── http_generic.rs            # `HttpGenericRules` — shared `TemplateProviderRules` for HTTP-transport providers (Ollama, MiniMax) and Qwen; `target_dir()=None`; `hook_gate` emits orchestrator instruction text (not a shell command); `skill_link` inlines full body; `http_generic_rules()` singleton helper  [Issue #705]
+│   │   │   ├── mod.rs                     # `TemplateProviderRules` trait (per-provider placeholder validation + unsupported-feature guard); `NullRules` fail-closed stub; `null_rules()` free fn returning `&'static NullRules`; re-exports `ClaudeRules`, `codex_rules()`, and `http_generic_rules()`; `read_sandboxed` + `read_skill_body` shared helpers (consolidates three near-identical per-file readers); Unix-gated `symlink_escape_is_rejected_by_starts_with_check` test  [Issue #701, #703, #704, #705, #727]
+│   │   │   ├── claude.rs                  # `ClaudeRules` — concrete `TemplateProviderRules` impl for Claude Code; `include()` delegates to `read_sandboxed`; 8 unit tests  [Issue #703, #727]
+│   │   │   ├── codex.rs                   # `CodexRules` — concrete `TemplateProviderRules` impl for Codex; `target_dir()=None`; `invoke_subagent` inlines as `## Sub-task`; `hook_gate` emits provider-neutral bash path; `include()` delegates to `read_sandboxed`; `skill_link` delegates to `read_skill_body`; `subagent_list` returns link-free Markdown table  [Issue #704, #727]
+│   │   │   └── http_generic.rs            # `HttpGenericRules` — shared `TemplateProviderRules` for HTTP-transport providers (Ollama, MiniMax) and Qwen; `target_dir()=None`; `hook_gate` emits orchestrator instruction text (not a shell command); `include()` delegates to `read_sandboxed`; `skill_link` delegates to `read_skill_body`; `http_generic_rules()` singleton helper  [Issue #705, #727]
 │   │   └── renderer/
 │   │       ├── mod.rs                     # Placeholder-expansion engine; `MAX_INCLUDE_DEPTH = 8`; `render_template()` entry point; include-cycle detection  [Issue #701]
 │   │       ├── tokenize.rs                # Hand-written tokenizer (no regex); splits template source into `Token::Literal` / `Token::Placeholder` / `Token::Include` variants; detects unterminated placeholder delimiters  [Issue #701]
@@ -764,10 +764,10 @@ maestro/
 | `src/templates/` | Canonical command-template render engine; `render_for_provider(provider, command)` is the public entry point; sandbox-validated path traversal; MAX_INCLUDE_DEPTH=8 include guard (Issue #701) |
 | `src/templates/error.rs` | `TemplateError` enum (thiserror); 11 variants covering all failure modes from manifest I/O to sandbox escape and include cycles (Issue #701) |
 | `src/templates/manifest.rs` | `Manifest::load(path)` — deserializes `.maestro/templates/manifest.toml`; `#[serde(deny_unknown_fields)]` enforces forward-compatibility (Issue #701) |
-| `src/templates/provider_rules/mod.rs` | `TemplateProviderRules` trait; `NullRules` fail-closed stub; `null_rules()` helper returns `&'static NullRules`; re-exports `ClaudeRules`, `codex_rules()`, and `http_generic_rules()` (Issues #701, #703, #704, #705) |
-| `src/templates/provider_rules/claude.rs` | `ClaudeRules` — concrete `TemplateProviderRules` for Claude Code; sandbox-validated `include()` guard; 11 unit tests (Issue #703) |
-| `src/templates/provider_rules/codex.rs` | `CodexRules` — concrete `TemplateProviderRules` for Codex; `target_dir()=None`; `invoke_subagent` inlines as `## Sub-task: {name}`; `hook_gate` emits provider-neutral `bash .maestro/hooks/<script>`; `skill_link` inlines full skill body; `subagent_list` returns link-free Markdown table (Issue #704) |
-| `src/templates/provider_rules/http_generic.rs` | `HttpGenericRules` — shared `TemplateProviderRules` for HTTP-transport providers (Ollama, MiniMax) and Qwen; `target_dir()=None`; `hook_gate` renders orchestrator instruction text instead of a shell command; `skill_link` inlines full skill body; `http_generic_rules()` singleton; 14 unit tests (Issue #705) |
+| `src/templates/provider_rules/mod.rs` | `TemplateProviderRules` trait; `NullRules` fail-closed stub; `null_rules()` helper returns `&'static NullRules`; re-exports `ClaudeRules`, `codex_rules()`, and `http_generic_rules()`; `read_sandboxed` + `read_skill_body` shared helpers consolidate the three formerly per-file sandbox-aware readers; Unix-gated `symlink_escape_is_rejected_by_starts_with_check` test (Issues #701, #703, #704, #705, #727) |
+| `src/templates/provider_rules/claude.rs` | `ClaudeRules` — concrete `TemplateProviderRules` for Claude Code; `include()` delegates to `read_sandboxed`; 8 unit tests (Issues #703, #727) |
+| `src/templates/provider_rules/codex.rs` | `CodexRules` — concrete `TemplateProviderRules` for Codex; `target_dir()=None`; `invoke_subagent` inlines as `## Sub-task: {name}`; `hook_gate` emits provider-neutral `bash .maestro/hooks/<script>`; `include()` delegates to `read_sandboxed`; `skill_link` delegates to `read_skill_body`; `subagent_list` returns link-free Markdown table (Issues #704, #727) |
+| `src/templates/provider_rules/http_generic.rs` | `HttpGenericRules` — shared `TemplateProviderRules` for HTTP-transport providers (Ollama, MiniMax) and Qwen; `target_dir()=None`; `hook_gate` renders orchestrator instruction text instead of a shell command; `include()` delegates to `read_sandboxed`; `skill_link` delegates to `read_skill_body`; `http_generic_rules()` singleton; 12 unit tests (Issues #705, #727) |
 | `src/templates/renderer/mod.rs` | Placeholder-expansion engine; walks token stream produced by `tokenize.rs`; enforces `MAX_INCLUDE_DEPTH = 8`; detects include cycles via a seen-set (Issue #701) |
 | `src/templates/renderer/tokenize.rs` | Hand-written tokenizer (no regex); emits `Token::Literal`, `Token::Placeholder`, `Token::Include`; returns `TemplateError::UnterminatedPlaceholder` on malformed input (Issue #701) |
 | `src/templates/renderer/tests/mod.rs` | 21 unit tests for the renderer: literal pass-through, placeholder expansion, unknown-placeholder error, include depth limit, sandbox escape rejection, cycle detection, and `NullRules` fail-closed (Issue #701) |
